renaming user_element.py to library.py
