+++ src/taipy/gui/extension/__init__.py
@@ -10,4 +10,4 @@
 # specific language governing permissions and limitations under the License.
 
 from ..types import PropertyType
-from .user_element import Element, ElementLibrary, ElementProperty
+from .library import Element, ElementLibrary, ElementProperty
